INT-191 introduce buildRepairCodeJobs
